// src/generated/config_sync/config_sync.ts
export class AvailableConfigsRequest extends pb_1.Message {
    #one_of_decls: number[][] = [];
    constructor(data?: any[] | {
        userId?: string;
    }) {
        super();
        pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
        if (!Array.isArray(data) && typeof data == "object") {
            if ("userId" in data && data.userId != undefined) {
                this.userId = data.userId;
            }
        }
    }
    get userId() {
        return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
    }
    set userId(value: string) {
        pb_1.Message.setField(this, 1, value);
    }
    static fromObject(data: {
        userId?: string;
    }): AvailableConfigsRequest {
        const message = new AvailableConfigsRequest({});
        if (data.userId != null) {
            message.userId = data.userId;
        }
        return message;
    }
    toObject() {
        const data: {
            userId?: string;
        } = {};
        if (this.userId != null) {
            data.userId = this.userId;
        }
        return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
        const writer = w || new pb_1.BinaryWriter();
        if (this.userId.length)
            writer.writeString(1, this.userId);
        if (!w)
            return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): AvailableConfigsRequest {
        const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new AvailableConfigsRequest();
        while (reader.nextField()) {
            if (reader.isEndGroup())
                break;
            switch (reader.getFieldNumber()) {
                case 1:
                    message.userId = reader.readString();
                    break;
                default: reader.skipField();
            }
        }
        return message;
    }
    serializeBinary(): Uint8Array {
        return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): AvailableConfigsRequest {
        return AvailableConfigsRequest.deserialize(bytes);
    }
}
export class AvailableConfigsResponse extends pb_1.Message {
    #one_of_decls: number[][] = [];
    constructor(data?: any[] | {
        configNames?: string[];
    }) {
        super();
        pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [1], this.#one_of_decls);
        if (!Array.isArray(data) && typeof data == "object") {
            if ("configNames" in data && data.configNames != undefined) {
                this.configNames = data.configNames;
            }
        }
    }
    get configNames() {
        return pb_1.Message.getFieldWithDefault(this, 1, []) as string[];
    }
    set configNames(value: string[]) {
        pb_1.Message.setField(this, 1, value);
    }
    static fromObject(data: {
        configNames?: string[];
    }): AvailableConfigsResponse {
        const message = new AvailableConfigsResponse({});
        if (data.configNames != null) {
            message.configNames = data.configNames;
        }
        return message;
    }
    toObject() {
        const data: {
            configNames?: string[];
        } = {};
        if (this.configNames != null) {
            data.configNames = this.configNames;
        }
        return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
        const writer = w || new pb_1.BinaryWriter();
        if (this.configNames.length)
            writer.writeRepeatedString(1, this.configNames);
        if (!w)
            return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): AvailableConfigsResponse {
        const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new AvailableConfigsResponse();
        while (reader.nextField()) {
            if (reader.isEndGroup())
                break;
            switch (reader.getFieldNumber()) {
                case 1:
                    pb_1.Message.addToRepeatedField(message, 1, reader.readString());
                    break;
                default: reader.skipField();
            }
        }
        return message;
    }
    serializeBinary(): Uint8Array {
        return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): AvailableConfigsResponse {
        return AvailableConfigsResponse.deserialize(bytes);
    }
}
export class ConfigSyncUpRequest extends pb_1.Message {
    #one_of_decls: number[][] = [];
    constructor(data?: any[] | {
        userId?: string;
        config?: Config;
    }) {
        super();
        pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
        if (!Array.isArray(data) && typeof data == "object") {
            if ("userId" in data && data.userId != undefined) {
                this.userId = data.userId;
            }
            if ("config" in data && data.config != undefined) {
                this.config = data.config;
            }
        }
    }
    get userId() {
        return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
    }
    set userId(value: string) {
        pb_1.Message.setField(this, 1, value);
    }
    get config() {
        return pb_1.Message.getWrapperField(this, Config, 3) as Config;
    }
    set config(value: Config) {
        pb_1.Message.setWrapperField(this, 3, value);
    }
    get has_config() {
        return pb_1.Message.getField(this, 3) != null;
    }
    static fromObject(data: {
        userId?: string;
        config?: ReturnType<typeof Config.prototype.toObject>;
    }): ConfigSyncUpRequest {
        const message = new ConfigSyncUpRequest({});
        if (data.userId != null) {
            message.userId = data.userId;
        }
        if (data.config != null) {
            message.config = Config.fromObject(data.config);
        }
        return message;
    }
    toObject() {
        const data: {
            userId?: string;
            config?: ReturnType<typeof Config.prototype.toObject>;
        } = {};
        if (this.userId != null) {
            data.userId = this.userId;
        }
        if (this.config != null) {
            data.config = this.config.toObject();
        }
        return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
        const writer = w || new pb_1.BinaryWriter();
        if (this.userId.length)
            writer.writeString(1, this.userId);
        if (this.has_config)
            writer.writeMessage(3, this.config, () => this.config.serialize(writer));
        if (!w)
            return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): ConfigSyncUpRequest {
        const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new ConfigSyncUpRequest();
        while (reader.nextField()) {
            if (reader.isEndGroup())
                break;
            switch (reader.getFieldNumber()) {
                case 1:
                    message.userId = reader.readString();
                    break;
                case 3:
                    reader.readMessage(message.config, () => message.config = Config.deserialize(reader));
                    break;
                default: reader.skipField();
            }
        }
        return message;
    }
    serializeBinary(): Uint8Array {
        return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): ConfigSyncUpRequest {
        return ConfigSyncUpRequest.deserialize(bytes);
    }
}
// ...
interface GrpcUnaryServiceInterface<P, R> {
    (message: P, metadata: grpc_1.Metadata, options: grpc_1.CallOptions, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
    (message: P, metadata: grpc_1.Metadata, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
    (message: P, options: grpc_1.CallOptions, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
    (message: P, callback: grpc_1.requestCallback<R>): grpc_1.ClientUnaryCall;
}
// ...
export class ConfigSyncClient extends grpc_1.makeGenericClientConstructor(UnimplementedConfigSyncService.definition, "ConfigSync", {}) {
    constructor(address: string, credentials: grpc_1.ChannelCredentials, options?: Partial<grpc_1.ChannelOptions>) {
        super(address, credentials, options);
    }
    GetAvailableConfigs: GrpcUnaryServiceInterface<AvailableConfigsRequest, AvailableConfigsResponse> = (message: AvailableConfigsRequest, metadata: grpc_1.Metadata | grpc_1.CallOptions | grpc_1.requestCallback<AvailableConfigsResponse>, options?: grpc_1.CallOptions | grpc_1.requestCallback<AvailableConfigsResponse>, callback?: grpc_1.requestCallback<AvailableConfigsResponse>): grpc_1.ClientUnaryCall => {
        return super.GetAvailableConfigs(message, metadata, options, callback);
    };
    ConfigSyncUp: GrpcUnaryServiceInterface<ConfigSyncUpRequest, ConfigSyncUpResponse> = (message: ConfigSyncUpRequest, metadata: grpc_1.Metadata | grpc_1.CallOptions | grpc_1.requestCallback<ConfigSyncUpResponse>, options?: grpc_1.CallOptions | grpc_1.requestCallback<ConfigSyncUpResponse>, callback?: grpc_1.requestCallback<ConfigSyncUpResponse>): grpc_1.ClientUnaryCall => {
        return super.ConfigSyncUp(message, metadata, options, callback);
    };
    ConfigSyncDown: GrpcUnaryServiceInterface<ConfigSyncDownRequest, Config> = (message: ConfigSyncDownRequest, metadata: grpc_1.Metadata | grpc_1.CallOptions | grpc_1.requestCallback<Config>, options?: grpc_1.CallOptions | grpc_1.requestCallback<Config>, callback?: grpc_1.requestCallback<Config>): grpc_1.ClientUnaryCall => {
        return super.ConfigSyncDown(message, metadata, options, callback);
    };
    ConfigCreateDelete: GrpcUnaryServiceInterface<ConfigCreateDeleteRequest, ConfigCreateDeleteResponse> = (message: ConfigCreateDeleteRequest, metadata: grpc_1.Metadata | grpc_1.CallOptions | grpc_1.requestCallback<ConfigCreateDeleteResponse>, options?: grpc_1.CallOptions | grpc_1.requestCallback<ConfigCreateDeleteResponse>, callback?: grpc_1.requestCallback<ConfigCreateDeleteResponse>): grpc_1.ClientUnaryCall => {
        return super.ConfigCreateDelete(message, metadata, options, callback);
    };
}